requirements.txt
alabaster==0.7.12
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
docutils==0.16
idna==2.10
imagesize==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
nmap==0.0.1
packaging==20.8
pkg-resources==0.0.0
Pygments==2.7.3
pyparsing==2.4.7
python3-nmap==1.4.9
pytz==2020.5
requests==2.25.1
rich==9.6.1
simplejson==3.17.2
snowballstemmer==2.0.0
Sphinx==3.4.2
sphinx-rtd-theme==0.5.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
synologydsm-api==1.0.1
tabulate==0.8.7
typing-extensions==3.7.4.3
urllib3==1.26.2
